zmiana

+++ Dzielniki - szkopul/Dzielniki/Dzielniki/Dzielniki.cpp
@@ -4,9 +4,9 @@
 
  int main()
  {
-	 int n;
+	 int64_t n;
 	 cin >> n;
-	 for (int i = 1; i <= n; i++)
+	 for (int64_t i = 1; i <= n; i++)
 		 if (n % i == 0) {
 			 cout << i << "\n";
